// src/types/define.ts
// 상수값들 정의 ******************************************************************************

export const SWIPE_THRESHOLD_RATIO = 0.12;
export const SWIPE_THRESHOLD_MIN_PX = 48;
export const AUTO_SLIDE_DELAY_MS = 5000;

// ******************************************************************************


export const enum ServicePlatform {
	APPLE_STORE = "apple-store",
  GOOGLE_PLAY = "google-play",
  STEAM = "steam",
  APP_STORE = "app-store",
  WINDOWS = "windows",
  MACOS = "macos",
  LINUX = "linux",
}